test: port the team statement suites to stmttest
Create, Get, UpdateTeam and Deactivate asserted the same behavior in both
dialect packages. Move them to stmttest so they run through
service.AllStatements once per dialect (ADR 041), as #664 did for the other
entities.

Two assertions tighten in the move: a duplicate (project_id, id) now expects
UniqueError instead of a bare error, and an over-length name expects
CheckError on both engines.

The empty-id case moves too. Both CreateTeam implementations reject it with
the same Go guard, so it never reaches the postgres CHECK (id <> '') that
Spanner lacks.

Postgres keeps the cascade test and the withTransaction rollback test, which
need raw SQL and a pgx wrapper.
